@@ -9,7 +9,7 @@ authors = ["n0 team"]
 repository = "https://github.com/n0-computer/iroh"
 
 # Sadly this also needs to be updated in .github/workflows/ci.yml
-rust-version = "1.70"
+rust-version = "1.71"
 
 [dependencies]
 anyhow = "1"
@@ -18,7 +18,7 @@ crossbeam = "0.8.2"
 data-encoding = "2.4.0"
 derive_more = { version = "1.0.0-beta.1", features = ["debug", "deref", "display", "from", "try_into", "into", "as_ref"] }
 ed25519-dalek = { version = "2.0.0", features = ["serde", "rand_core"] }
-flume = "0.10"
+flume = "0.11"
 iroh-bytes = { version = "0.7.0", path = "../iroh-bytes" }
 iroh-metrics = { version = "0.7.0", path = "../iroh-metrics", optional = true }
 once_cell = "1.18.0"
@@ -26,6 +26,7 @@ postcard = { version = "1", default-features = false, features = ["alloc", "use-
 rand = "0.8.5"
 rand_core = "0.6.4"
 serde = { version = "1.0.164", features = ["derive"] }
+strum = { version = "0.25", features = ["derive"] }
 url = "2.4"
 bytes = "1"
 parking_lot = "0.12.1"
