@@ -521,7 +521,6 @@ impl DocCommands {
                         LiveEvent::NeighborDown(peer) => {
                             println!("neighbor peer down: {peer:?}");
                         }
-                        LiveEvent::Closed => println!("document closed"),
                     }
                 }
             }
@@ -550,7 +549,7 @@ impl DocCommands {
 
 async fn get_doc(iroh: &Iroh, env: &ConsoleEnv, id: Option<NamespaceId>) -> anyhow::Result<Doc> {
     iroh.docs
-        .get(env.doc(id)?)
+        .open(env.doc(id)?)
         .await?
         .context("Document not found")
 }
